Added screenshot tests and tests for missing elements.
