+++ tests/js/takeScreenshot/html/takeScreenshot.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Screenshot Demo</title>
+</head>
+<body>
+    <svg width="400" height="400">
+        <polygon points="10,10 200,190 390,10 210,200 390,390, 200,210 10,390 190,200 10,10 " style="fill:lime;stroke:purple;stroke-width:5"></polygon>
+        A polygon shaped like a stylized "X".
+    </svg>
+</body>
+</html>
